works in RootCore (xAODTrigL1Calo-00-00-38)
	* RootCore compatibility
	* Tag as xAODTrigL1Calo-00-00-38

2015-02-11 John Morris <john.morris@cern.ch>
	* remove layer as data member of TriggerTowerAuxContainer_v2
	* save 7168 bytes per event
	* Tag as xAODTrigL1Calo-00-00-37

2015-01-31 John Morris <john.morris@Cern.ch>
	* Lots of initialize functions for all v2 classes
	* Tag as xAODTrigL1Calo-00-00-36

2015-01-30 John Morris <john.morris@cern.ch>
        * Optimise type to minimise disk usage
	* CMXCPTob_v2
	* CMXJetHits_v2
	* CMXEtSums_v2

2015-01-30 John Morris <john.morris@cern.ch>
	* Optimise type to minimise disk usage
...
(Long ChangeLog diff - truncated)

+++ Event/xAOD/xAODTrigL1Calo/xAODTrigL1Calo/versions/CMXCPTobAuxContainer_v2.h
@@ -4,7 +4,7 @@
   Copyright (C) 2002-2017 CERN for the benefit of the ATLAS collaboration
 */
 
-// $Id: CMXCPTobAuxContainer_v2.h 642659 2015-01-29 12:41:06Z morrisj $
+// $Id: CMXCPTobAuxContainer_v2.h 646335 2015-02-12 01:16:10Z morrisj $
 #ifndef XAODTRIGL1CALO_VERSIONS_CMXCPTOBAUXCONTAINER_V2_H
 #define XAODTRIGL1CALO_VERSIONS_CMXCPTOBAUXCONTAINER_V2_H
 
@@ -22,8 +22,8 @@ namespace xAOD{
   ///
   /// @author John Morris <john.morris@cern.ch>
   ///
-  /// $Revision: 642659 $
-  /// $Date: 2015-01-29 13:41:06 +0100 (Thu, 29 Jan 2015) $
+  /// $Revision: 646335 $
+  /// $Date: 2015-02-12 02:16:10 +0100 (Thu, 12 Feb 2015) $
 
   class CMXCPTobAuxContainer_v2 : public AuxContainerBase {
     public:
@@ -31,16 +31,16 @@ namespace xAOD{
       CMXCPTobAuxContainer_v2();
 
     private:
-      std::vector<int> crate;
-      std::vector<int> cmx;
-      std::vector<int> cpm;
-      std::vector<int> chip;
-      std::vector<int> location;
+      std::vector<uint_least8_t> crate;
+      std::vector<uint_least8_t> cmx;
+      std::vector<uint_least8_t> cpm;
+      std::vector<uint_least8_t> chip;
+      std::vector<uint_least8_t> location;
       std::vector<uint_least8_t> peak;
-      std::vector<std::vector<int>> energyVec;
-      std::vector<std::vector<int>> isolationVec;
-      std::vector<std::vector<int>> errorVec;
-      std::vector<std::vector<unsigned int>> presenceMapVec;
+      std::vector<std::vector<uint_least8_t> > energyVec;
+      std::vector<std::vector<uint_least8_t> > isolationVec;
+      std::vector<std::vector<uint_least8_t> > errorVec;
+      std::vector<std::vector<uint_least16_t> > presenceMapVec;
 
   }; // class CMXCPTobAuxContainer_v2
 } // namespace xAOD
