+++ Event/xAOD/xAODTrigL1Calo/xAODTrigL1Calo/versions/CMXEtSumsAuxContainer_v2.h
@@ -4,7 +4,7 @@
   Copyright (C) 2002-2017 CERN for the benefit of the ATLAS collaboration
 */
 
-// $Id: CMXEtSumsAuxContainer_v2.h 642659 2015-01-29 12:41:06Z morrisj $
+// $Id: CMXEtSumsAuxContainer_v2.h 646335 2015-02-12 01:16:10Z morrisj $
 #ifndef XAODTRIGL1CALO_VERSIONS_CMXETSUMSAUXCONTAINER_V2_H
 #define XAODTRIGL1CALO_VERSIONS_CMXETSUMSAUXCONTAINER_V2_H
 
@@ -22,8 +22,8 @@ namespace xAOD{
   ///
   /// @author John Morris <john.morris@cern.ch>
   ///
-  /// $Revision: 642659 $
-  /// $Date: 2015-01-29 13:41:06 +0100 (Thu, 29 Jan 2015) $
+  /// $Revision: 646335 $
+  /// $Date: 2015-02-12 02:16:10 +0100 (Thu, 12 Feb 2015) $
 
   class CMXEtSumsAuxContainer_v2 : public AuxContainerBase {
     public:
@@ -31,15 +31,15 @@ namespace xAOD{
       CMXEtSumsAuxContainer_v2();
 
     private:
-      std::vector<int> crate;
-      std::vector<int> source;
+      std::vector<uint_least8_t> crate;
+      std::vector<uint_least8_t> source;
       std::vector<uint_least8_t> peak;
-      std::vector<std::vector<unsigned int>> etVec;
-      std::vector<std::vector<unsigned int>> exVec;
-      std::vector<std::vector<unsigned int>> eyVec;
-      std::vector<std::vector<int>> etErrorVec;
-      std::vector<std::vector<int>> exErrorVec;
-      std::vector<std::vector<int>> eyErrorVec;
+      std::vector<std::vector<uint_least16_t> > etVec;
+      std::vector<std::vector<uint_least16_t> > exVec;
+      std::vector<std::vector<uint_least16_t> > eyVec;
+      std::vector<std::vector<uint_least8_t> > etErrorVec;
+      std::vector<std::vector<uint_least8_t> > exErrorVec;
+      std::vector<std::vector<uint_least8_t> > eyErrorVec;
 
   }; // class CMXEtSumsAuxContainer_v2
 } // namespace xAOD
